twitter-bot/requirements.txt
tweepy==4.14.0
anthropic==0.42.0
python-dotenv==1.0.0
httpx==0.26.0
schedule==1.2.1
